requirements.txt
numpy-1.18
pandas-1.0.3
python-dateutil-2.8.1
pytz-2020.1
joblib-0.14.1
scikit-learn-0.22.2
scipy-1.4.1


